se agrego la funcionalidad de logout
se implemento la logica para el logout de la tarea 140
