+++ src/Components/Auth/RegisterForm.jsx
@@ -1,48 +1,43 @@
-import React, { useEffect } from "react";
+/* eslint-disable no-undef */
+import React from "react";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import * as yup from "yup";
+import "../FormStyles.css";
+import { Link, useNavigate } from "react-router-dom";
 import { yupErrorMessages, yupRegexValidation } from "utils/messages/formMessagesValidation";
-import blogImg02 from "Assets/images/blog-img-02.jpg";
-import { signIn } from "store/Slices/authSlice";
-import { Navigate, useNavigate, Link } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import { clearMessage } from "store/Slices/messageSlice";
-import { error } from "utils/alerts/alerts";
+import somosmas from "Assets/images/LOGO-SOMOSMAS.png";
+import imgRegister from "Assets/images/image-loginRegistrer.jpg";
+import { useDispatch } from "react-redux";
+import { signUp } from "store/Slices/authSlice";
 
-const LoginForm = () => {
-	const dispatch = useDispatch();
+const RegisterForm = () => {
+	// eslint-disable-next-line no-unused-vars
 	const navigate = useNavigate();
+	const dispatch = useDispatch();
 
-	const { isLoggedIn } = useSelector(state => state.user);
-	useEffect(() => {
-		dispatch(clearMessage());
-	}, [dispatch]);
-
-	if (isLoggedIn) {
-		return <Navigate to="/backoffice" />;
-	}
 	return (
 		<div className="flex w-full bg-slate-50  justify-between items-center min-h-screen">
 			<div className="w-full sm:w-full sm:mx-auto md:w-1/2 md:mx-auto flex flex-col justify-center items-center">
 				<Formik
-					initialValues={{ email: "", password: "" }}
+					initialValues={{ name: "", email: "", password: "", confirmPassword: "" }}
 					onSubmit={(values, { resetForm }) => {
-						dispatch(signIn({ email: values.email, password: values.password }))
-							.unwrap()
-							.then(() => {
-								navigate("/backoffice");
-							})
-							.catch(() => {
-								error();
-							});
+						dispatch(signUp({ name: values.name, email: values.email, password: values.password }));
+						navigate("/login-user");
+						resetForm(values);
 					}}
 					validationSchema={() =>
 						yup.object().shape({
+							name: yup.string().required(yupErrorMessages.required),
 							email: yup.string().email(yupErrorMessages.invalidEmail).required(yupErrorMessages.required),
 							password: yup
 								.string()
 								.matches(yupRegexValidation.messageRgx, yupErrorMessages.password6)
 								.required(yupErrorMessages.required),
+							confirmPassword: yup
+
+								.string()
+								.required(yupErrorMessages.required)
+								.oneOf([yup.ref("password"), null], yupErrorMessages.comparePass),
 						})
 					}
 				>
@@ -51,15 +46,26 @@ const LoginForm = () => {
 							<div className="w-full flex flex-col  gap-4">
 								<div className="hidden lg:block md:hidden sm:hidden">
 									<h4 className="text-base text-left">Bienvenido</h4>
-									<h1 className="sefl-start text-2xl md:text-3xl text-left font-semibold">¡Ingresá a tu cuenta!</h1>
+									<h1 className="sefl-start text-xl md:text-3xl text-left font-semibold">
+										Ingresa tus datos de registro!
+									</h1>
 								</div>
 								<div className="mx-auto lg:hidden md:block">
-									<img src="images/logo-somosmas.png" />
+									<img src={somosmas} />
 								</div>
 								<Field
 									className="h-14 w-full border border-slate-300 rounded-lg p-4"
+									name="name"
+									placeholder="Ingresa tu nombre"
+								/>
+								<ErrorMessage
+									name="name"
+									component={() => <span className="text-red-400 text-xs">{errors.name}</span>}
+								/>
+								<Field
+									className="h-14 w-full border border-slate-300 rounded-lg p-4"
 									name="email"
-									placeholder="Ingresa tu correo electrónico"
+									placeholder="Ingresa tu correo electronico"
 								/>
 								<ErrorMessage
 									name="email"
@@ -70,11 +76,22 @@ const LoginForm = () => {
 									className="h-14 w-full border border-slate-300 rounded-lg p-4"
 									type="password"
 									name="password"
-									placeholder="Ingresa tu contraseña"
+									placeholder="Ingresa una contraseña nueva"
 								/>
 								<ErrorMessage
 									name="password"
 									component={() => <span className="text-red-400 text-xs">{errors.password}</span>}
+								/>
+
+								<Field
+									className="h-14 w-full border border-slate-300 rounded-lg p-4"
+									name="confirmPassword"
+									type="password"
+									placeholder="Ingresa nuevamente la contraseña"
+								/>
+								<ErrorMessage
+									name="confirmPassword"
+									component={() => <span className="text-red-400 text-xs">{errors.confirmPassword}</span>}
 								/>
 
 								<button
@@ -83,25 +100,25 @@ const LoginForm = () => {
 									rounded-lg  mx-auto hover:bg-red-500 hover:-translate-y-1 
 									transition-all duration-500 text-white text-xl font-medium"
 								>
-									Ingresá
+									Registrarme
 								</button>
 							</div>
 						</Form>
 					)}
 				</Formik>
 				<div className="absolute bottom-4 flex gap-2">
-					<p className="font-medium text-slate-600">¿No tienes cuenta?</p>
-					<Link to="/register-user" className="text-red-600 font-medium">
-						Registrarte
+					<p className="font-medium text-slate-600">Ya tienes cuenta?</p>
+					<Link to="/login-user" className="text-red-600 font-medium">
+						Inicia sesion
 					</Link>
 				</div>
 			</div>
 
 			<div className="hidden lg:w-1/2 lg:block h-screen md:w-1/2 md:hidden sm:hidden">
-				<img alt="loginRegister" src={blogImg02} className="h-screen w-full imgLogin" />
+				<img alt="loginRegister" src={imgRegister} className="h-screen w-full" />
 			</div>
 		</div>
 	);
 };
 
-export default LoginForm;
+export default RegisterForm;
