+++ src/Components/Home/index.jsx
@@ -4,8 +4,15 @@ import Slider from "./Slider";
 import Staff from "./Staff";
 import Testimonials from "./Testimonials";
 import WelcomeText from "./WelcomeText";
+import { Navigate } from "react-router-dom";
+import { useSelector } from "react-redux";
 
 export const Home = () => {
+	const { isLoggedIn } = useSelector(state => state.user);
+
+	if (!isLoggedIn) {
+		return <Navigate to="/login-user" />;
+	}
 	return (
 		<div className=" flex flex-col w-full">
 			<div className=" flex flex-col justify-start align-items-center">
