+++ src/Services/UsersAdmin/ApiService.js
@@ -1,5 +1,8 @@
 import axios from "axios";
-import { success, erase, error, update } from "utils/alerts/alerts";
+import { success, erase, error } from "utils/alerts/alerts";
+import { Put } from "Services/privateApiService";
+
+const usersEndPoint = "/users";
 
 export const getUserAdmin = (setUser, id) => {
 	axios
@@ -13,19 +16,13 @@ export const getUserAdmin = (setUser, id) => {
 		});
 };
 
-export const getUsersAdmin = (
-	setUser,
-	amountToShow,
-	page,
-	filterTypeOfUser,
-	inputFilter
-) => {
+export const getUsersAdmin = (setUser, amountToShow, page, filterTypeOfUser, inputFilter) => {
 	axios
 		.get(
 			"https://ongapi.alkemy.org/api/" +
-				`users?limit=${amountToShow}&skip=${amountToShow * page}${
-					filterTypeOfUser && "&role=" + filterTypeOfUser
-				}${inputFilter.length>=2 && "&search=" + inputFilter}`
+				`users?limit=${amountToShow}&skip=${amountToShow * page}${filterTypeOfUser && "&role=" + filterTypeOfUser}${
+					inputFilter.length >= 2 && "&search=" + inputFilter
+				}`
 		)
 		.then(res => {
 			console.log(res.data.data);
@@ -36,17 +33,13 @@ export const getUsersAdmin = (
 			console.log(err);
 		});
 };
-export const getAmountOfUsersAdmin = (
-	setAmountOfUsers,
-	filterTypeOfUser,
-	inputFilter
-) => {
+export const getAmountOfUsersAdmin = (setAmountOfUsers, filterTypeOfUser, inputFilter) => {
 	axios
 		.get(
 			"https://ongapi.alkemy.org/api/" +
-				`users${filterTypeOfUser && "?role=" + filterTypeOfUser}${
-					filterTypeOfUser ? "&" : "?"
-				}${inputFilter.length>=2 && "search=" + inputFilter}`
+				`users${filterTypeOfUser && "?role=" + filterTypeOfUser}${filterTypeOfUser ? "&" : "?"}${
+					inputFilter.length >= 2 && "search=" + inputFilter
+				}`
 		)
 		.then(res => {
 			setAmountOfUsers(res.data.data.length);
@@ -57,20 +50,8 @@ export const getAmountOfUsersAdmin = (
 		});
 };
 
-export const putUserAdmin = (id, values) => {
-	axios
-		.put(`https://ongapi.alkemy.org/api/users/${id}`, {
-			...values,
-			group_id: 4,
-		})
-		.then(res => {
-			update();
-			console.log(res);
-		})
-		.catch(err => {
-			error();
-			console.log(err);
-		});
+export const putUserAdmin = async (id, values) => {
+	await Put(`${usersEndPoint}/${id}`, values);
 };
 
 export const postUserAdmin = values => {
